+++ apps/chronicles/navigation/index.tsx
@@ -32,6 +32,7 @@ import { AccordionScreen } from "../screens/components/AccordionScreen";
 import { MenuScreen } from "../screens/components/MenuScreen";
 import { ProgressIndicatorScreen } from "../screens/components/ProgressIndicatorScreen";
 import { PortalScreen } from "../screens/components/PortalScreen";
+import { DialogScreen } from "../screens/components/DialogScreen";
 import { EnvironmentScreen } from "../screens/components/EnvironmentScreen";
 
 export default function Navigation({
@@ -111,9 +112,9 @@ function DiscoverNavigator() {
             <DiscoverStack.Screen name="Menu" component={MenuScreen} />
             <DiscoverStack.Screen name="ProgressIndicator" options={{ title: "Progress Indicators" }} component={ProgressIndicatorScreen} />
             <DiscoverStack.Screen name="Portal" component={PortalScreen} />
+            <DiscoverStack.Screen name="Dialog" component={DialogScreen} />
             <DiscoverStack.Screen name="Environment" component={EnvironmentScreen} />
-
-        </DiscoverStack.Navigator>
+        </DiscoverStack.Navigator >
     );
 }