+++ packages/components/src/components/EDSProvider/EDSProvider.tsx
@@ -2,7 +2,9 @@ import React, { createContext, PropsWithChildren } from "react";
 import { ColorScheme, Density } from "../../styling/types";
 import { PortalProvider } from "../Portal/PortalContext";
 import { Portal } from "../Portal";
+import { DialogServiceProvider } from "../Dialog/service/DialogServiceProvider";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { ScrimProvider } from "../_internal/ScrimProvider";
 
 export type EDSProviderProps = {
     colorScheme: ColorScheme;
@@ -23,12 +25,15 @@ export const EDSProvider = (
             value={{ colorScheme: props.colorScheme, density: props.density }}>
             <GestureHandlerRootView style={{ flex: 1 }}>
                 <PortalProvider>
-                    <Portal.Host style={{ flex: 1 }} name="root">
-                        {props.children}
-                    </Portal.Host>
+                    <ScrimProvider>
+                        <Portal.Host style={{ flex: 1 }} name="root">
+                            {props.children}
+                            <DialogServiceProvider />
+                        </Portal.Host>
+                    </ScrimProvider>
                 </PortalProvider>
             </GestureHandlerRootView>
-        </EDSContext.Provider>
+        </EDSContext.Provider >
     );
 };