+++ packages/components/src/styling/types.ts
@@ -62,6 +62,7 @@ export type MasterToken = {
                 sticky: ColorSchemeValues<Color>;
                 temporaryNav: ColorSchemeValues<Color>;
             };
+            scrim: ColorSchemeValues<string>
         };
         interactive: {
             primary: ColorSchemeValues<Color>;
@@ -119,6 +120,13 @@ export type MasterToken = {
                     widthLarge: DensityValues<number>;
                 };
             };
+            dialog: {
+                minHeight: DensityValues<number>,
+                defaultWidth: DensityValues<number>
+                header: {
+                    height: DensityValues<number>
+                }
+            }
         };
         shadow: {
             [TElev in Elevation]: ShadowStyle;
@@ -149,6 +157,10 @@ export type MasterToken = {
                 paddingVertical: DensityValues<number>;
                 iconGap: DensityValues<number>;
             }
+        },
+        dialog: {
+            padding: DensityValues<number>,
+            gap: DensityValues<number>
         }
         cell: {
             group: {